requirements.txt
python-telegram-bot==21.5
openai-whisper==20231117
ffmpeg-python==0.2.0
